% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reform.R
\name{apply_reforms}
\alias{apply_reforms}
\title{Apply Multiple Reforms to Assumptions}
\usage{
apply_reforms(assumptions, reforms, check_exclusivity = TRUE)
}
\arguments{
\item{assumptions}{Data frame of Social Security assumptions (e.g., tr2025)}

\item{reforms}{A list of Reform objects to apply}

\item{check_exclusivity}{Logical, whether to check for mutual exclusivity
conflicts. Default is TRUE.}
}
\value{
Modified assumptions data frame with all reforms applied
}
\description{
Applies a list of reforms to assumptions, checking for mutual exclusivity
conflicts. Reforms are applied in order.
}
\details{
When \code{check_exclusivity = TRUE} (default), the function will stop with
an error if any mutually exclusive reforms are detected. The mutual
exclusivity groups are:
\itemize{
\item PIA Formula reforms (#2-4)
\item NRA reforms (#5-7)
\item COLA Index reforms (#8-10)
\item Taxmax reforms (#12-14)
}
}
\examples{
\dontrun{
reforms <- list(
  reform_reduce_benefits(multiplier = 0.95, effective_year = 2030),
  reform_chained_cpi(effective_year = 2030),
  reform_nra_to_68(effective_year = 2030)
)
reformed <- apply_reforms(tr2025, reforms)
}

}
