% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reform_functions.R
\name{basic_minimum_benefit}
\alias{basic_minimum_benefit}
\title{Basic Minimum Benefit Calculation}
\usage{
basic_minimum_benefit(worker, assumptions, debugg = FALSE)
}
\arguments{
\item{worker}{Dataframe with a worker's benefits by age (must include wrk_ben, nra_ind)}

\item{assumptions}{Dataframe with the Social Security Trustees assumptions
(must include bmb_individual, bmb_couple if reform is active)}

\item{debugg}{Boolean value that directs function to output additional variables if set to true}
}
\value{
worker Dataframe with BMB supplement added to wrk_ben
}
\description{
Applies the Basic Minimum Benefit (BMB) reform that supplements low benefits
at or after Normal Retirement Age. The BMB is calculated as:
BMB_supplement = max(BMB_rate - 0.70 * actuarially_adjusted_benefit, 0)
}
\details{
This function should be called AFTER worker_benefit() in the calculation pipeline.
}
