% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benefit_calculations.R
\name{cola}
\alias{cola}
\title{COLA Calculation}
\usage{
cola(worker, assumptions, debugg = FALSE)
}
\arguments{
\item{worker}{Dataframe with a worker's unadjusted PIA -- both retired worker and spousal -- by age}

\item{assumptions}{Dataframe with the Social Security Trustees historical and projected economic variables and program parameters}

\item{debugg}{Boolean value that directs function to output additional variables if set to true}
}
\value{
worker Dataframe with a worker's COLA-adjusted retired worker and spousal PIA by age
}
\description{
Function that computes a worker's COLA-adjusted PIA
}
