% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eligibility.R
\name{comp_period}
\alias{comp_period}
\title{Computation Period}
\usage{
comp_period(worker, debugg = FALSE)
}
\arguments{
\item{worker}{Data frame with workers by age.}

\item{debugg}{Optional boolean variable that allows for the output of additional variables for testing.}
}
\value{
worker Data frame with workers by age, with computation period appended.
}
\description{
Function for determining a worker's computation period.
}
