% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earnings.R
\name{earnings_generator}
\alias{earnings_generator}
\title{Earnings Generator}
\usage{
earnings_generator(
  birth_yr = 1960,
  type = "medium",
  age_claim,
  age_elig = 62,
  factors,
  assumptions,
  zero_yrs = 0,
  custom_avg_earnings = NULL,
  age_stop = 65,
  spouse = FALSE
)
}
\arguments{
\item{birth_yr}{Numeric value representing the birth year of the worker.}

\item{type}{Character value specifying the earnings level of the worker.}

\item{age_claim}{Numeric value for the age in which the worker claims benefits.}

\item{age_elig}{Numeric value for the age in which the worker becomes eligible for benefits.}

\item{factors}{Data frame for the Trustees' scaled earnings factors.}

\item{assumptions}{Data frame of the pre-prepared Trustees assumptions.}

\item{zero_yrs}{Numeric value for the number of zero-earning years held by the worker.}

\item{custom_avg_earnings}{Numeric value for the real average earnings for the worker, if type="custom" is selected.}

\item{spouse}{Boolean value for whether the worker has a spouse -- not currently in use.}
}
\value{
worker Data frame with the earnings of the worker.
}
\description{
Function that generates the lifetime earnings for a specified worker
}
\examples{
\dontrun{
med_worker <- earnings_generator(birth_yr=1960, type="medium", age_claim = 67, factors=sef, assumptions=tr2025)
}

}
