% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benefit_calculations.R
\name{floor_dime}
\alias{floor_dime}
\title{Floor to Next Lower Dime}
\usage{
floor_dime(x)
}
\arguments{
\item{x}{Numeric value(s) representing dollar amounts}
}
\value{
Numeric value(s) rounded down to the nearest $0.10
}
\description{
Rounds a dollar amount down to the next lower multiple of $0.10, per
42 USC 415(a)(2)(C) (PIA rounding) and 42 USC 415(i)(2)(A)(ii) (COLA rounding).
}
\examples{
floor_dime(1234.56)  # Returns 1234.50
floor_dime(1234.99)  # Returns 1234.90

}
\keyword{internal}
