% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benefit_calculations.R
\name{generate_spouse_dependent_benefit}
\alias{generate_spouse_dependent_benefit}
\title{Generate Spouse's Dependent Benefit Based on Worker's Record (Internal)}
\usage{
generate_spouse_dependent_benefit(
  worker_data,
  spouse_spec,
  factors,
  assumptions
)
}
\arguments{
\item{worker_data}{Dataframe for a single worker containing year, age, basic_pia, cola_basic_pia}

\item{spouse_spec}{Character string with spouse specification}

\item{factors}{Data frame for the Trustees' scaled earnings factors}

\item{assumptions}{Data frame of the pre-prepared Trustees assumptions}
}
\value{
Numeric value of spouse's spousal benefit based on worker's record for this year
}
\description{
Calculates the spouse's spousal benefit that is based on the worker's earnings record.
This is used for RET calculations to determine the total benefit pot subject to reduction.
}
\keyword{internal}
