% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reform.R
\name{print.Reform}
\alias{print.Reform}
\title{Print Method for Reform Objects}
\usage{
\method{print}{Reform}(x, ...)
}
\arguments{
\item{x}{A Reform object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the Reform object
}
\description{
Print Method for Reform Objects
}
