% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eligibility.R
\name{qc_comp}
\alias{qc_comp}
\title{Quarters of Coverage Computation}
\usage{
qc_comp(worker, debugg = FALSE)
}
\arguments{
\item{worker}{Data frame with workers' earnings and QC requirements by age.}

\item{debugg}{Optional boolean variable that allows for the output of additional variables for testing.}
}
\value{
worker Data frame with a workers' earnings and QC's by age.
}
\description{
Function for computation a worker's QC's at each age. QC's are used for
determining eligibility.
}
