% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reform_templates.R
\name{reform_40_year_averaging}
\alias{reform_40_year_averaging}
\title{Create Reform: 40-Year Averaging Period}
\usage{
reform_40_year_averaging(effective_year = 2026)
}
\arguments{
\item{effective_year}{Year when the reform takes effect. Default is 2026.}
}
\value{
A Reform object
}
\description{
Creates a reform that changes the AIME computation period from 35 years
(with 5 dropout years) to 40 years (with 0 dropout years).
}
\details{
Phase-in: Reduce dropout years by 1 per year (5->4->3->2->1->0).
}
\examples{
\dontrun{
reform <- reform_40_year_averaging(effective_year = 2030)
}

}
