% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reform_templates.R
\name{reform_chained_cpi}
\alias{reform_chained_cpi}
\title{Create Reform: Index COLAs to Chained CPI}
\usage{
reform_chained_cpi(effective_year = 2026)
}
\arguments{
\item{effective_year}{Year when the reform takes effect. Default is 2026.}
}
\value{
A Reform object
}
\description{
Creates a reform that indexes COLAs to the Chained CPI-U (C-CPI-U) instead
of CPI-W. C-CPI-U grows about 0.3 percentage points slower than CPI-W.
}
\examples{
\dontrun{
reform <- reform_chained_cpi(effective_year = 2030)
}

}
