% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reform_templates.R
\name{reform_change_tax_rate}
\alias{reform_change_tax_rate}
\title{Create Reform: Change Payroll Tax Rate}
\usage{
reform_change_tax_rate(rate_change, effective_year = 2026)
}
\arguments{
\item{rate_change}{Change in tax rate in percentage points. E.g., 1.0 for +1\%.}

\item{effective_year}{Year when the reform takes effect. Default is 2026.}
}
\value{
A Reform object
}
\description{
Creates a reform that changes the OASI payroll tax rate by a specified
amount in percentage points.
}
\examples{
\dontrun{
# Increase payroll tax rate by 1 percentage point
reform <- reform_change_tax_rate(rate_change = 1.0, effective_year = 2030)
}

}
