% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reform_templates.R
\name{reform_cpi_e}
\alias{reform_cpi_e}
\title{Create Reform: Index COLAs to CPI-E}
\usage{
reform_cpi_e(effective_year = 2026)
}
\arguments{
\item{effective_year}{Year when the reform takes effect. Default is 2026.}
}
\value{
A Reform object
}
\description{
Creates a reform that indexes COLAs to the CPI for the Elderly (CPI-E).
CPI-E grows about 0.2 percentage points faster than CPI-W due to higher
health care spending by seniors.
}
\examples{
\dontrun{
reform <- reform_cpi_e(effective_year = 2030)
}

}
