% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reform_templates.R
\name{reform_eliminate_taxmax}
\alias{reform_eliminate_taxmax}
\title{Create Reform: Eliminate Taxmax with 15\% Benefit Credit}
\usage{
reform_eliminate_taxmax(effective_year = 2026)
}
\arguments{
\item{effective_year}{Year when the reform takes effect. Default is 2026.}
}
\value{
A Reform object
}
\description{
Creates a reform that eliminates the taxable maximum entirely and adds
a fourth PIA bracket with 15\% replacement rate for all earnings above
the old taxmax.
}
\examples{
\dontrun{
reform <- reform_eliminate_taxmax(effective_year = 2030)
}

}
