% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reform_templates.R
\name{reform_index_nra}
\alias{reform_index_nra}
\title{Create Reform: Index NRA to Longevity}
\usage{
reform_index_nra(effective_year = 2026)
}
\arguments{
\item{effective_year}{Year when the reform takes effect. Default is 2026.}
}
\value{
A Reform object
}
\description{
Creates a reform that indexes the Normal Retirement Age to life expectancy
improvements. Phase-in: +1 month per 2 eligibility cohorts, no cap.
}
\examples{
\dontrun{
reform <- reform_index_nra(effective_year = 2030)
}

}
