% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reform_templates.R
\name{reform_phase_out_spousal}
\alias{reform_phase_out_spousal}
\title{Create Reform: Phase Out Spousal Benefits}
\usage{
reform_phase_out_spousal(effective_year = 2026, phase_in_years = 10)
}
\arguments{
\item{effective_year}{Year when the reform takes effect. Default is 2026.}

\item{phase_in_years}{Number of years to phase out. Default is 10.}
}
\value{
A Reform object
}
\description{
Creates a reform that phases out spousal benefits by reducing the spousal
PIA share from 50\% to 0\%.
}
\examples{
\dontrun{
reform <- reform_phase_out_spousal(effective_year = 2030, phase_in_years = 20)
}

}
