% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reform.R
\name{reform_raise_nra}
\alias{reform_raise_nra}
\title{Create Common Reform: Raise Normal Retirement Age}
\usage{
reform_raise_nra(target_nra, effective_year, phase_in_years = 10)
}
\arguments{
\item{target_nra}{Target NRA (e.g., 69)}

\item{effective_year}{Year reform begins}

\item{phase_in_years}{Years to phase in (default 10)}
}
\value{
A Reform object
}
\description{
Creates a reform that raises the Normal Retirement Age.
}
\examples{
\dontrun{
reform <- reform_raise_nra(target_nra = 69, effective_year = 2030)
}

}
