% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reform_templates.R
\name{reform_reduce_fact3}
\alias{reform_reduce_fact3}
\title{Create Reform: Reduce Third Replacement Factor}
\usage{
reform_reduce_fact3(target_fact3 = 0.05, effective_year, phase_in_years = 10)
}
\arguments{
\item{target_fact3}{Target value for fact3. Default is 0.05 (5\%).}

\item{effective_year}{Year when the reform takes effect.}

\item{phase_in_years}{Number of years to phase in. Default is 10.}
}
\value{
A Reform object
}
\description{
Creates a reform that reduces the third PIA replacement factor (fact3) from
15\% to a target value. This reduces benefits for high earners.
}
\details{
Wrapper around reform_benefit_formula() for convenience.
}
\examples{
\dontrun{
# Reduce fact3 to 5\% over 10 years starting in 2030
reform <- reform_reduce_fact3(target_fact3 = 0.05, effective_year = 2030)
}

}
