% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reform_templates.R
\name{reform_repeal_ret}
\alias{reform_repeal_ret}
\title{Create Reform: Repeal Retirement Earnings Test}
\usage{
reform_repeal_ret(effective_year = 2026)
}
\arguments{
\item{effective_year}{Year when the reform takes effect. Default is 2026.}
}
\value{
A Reform object
}
\description{
Creates a reform that repeals the Retirement Earnings Test (RET), allowing
beneficiaries to receive full benefits regardless of earnings.
}
\examples{
\dontrun{
reform <- reform_repeal_ret(effective_year = 2030)
}

}
