% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reform_templates.R
\name{reform_simpson_bowles}
\alias{reform_simpson_bowles}
\title{Create Reform: Simpson-Bowles 4-Bracket PIA}
\usage{
reform_simpson_bowles(effective_year = 2026, phase_in_years = 10)
}
\arguments{
\item{effective_year}{Year when the reform takes effect. Default is 2026.}

\item{phase_in_years}{Number of years to phase in. Default is 10.}
}
\value{
A Reform object
}
\description{
Creates a reform that adds a fourth PIA bracket at AWI and modifies
replacement factors: fact2: 32\%->30\%, fact3: 15\%->10\%, fact4: 15\%->5\%.
Phased in over 10 eligibility cohorts.
}
\examples{
\dontrun{
reform <- reform_simpson_bowles(effective_year = 2030)
}

}
