% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benefit_calculations.R
\name{rf_and_drc}
\alias{rf_and_drc}
\title{Early Retirement Reduction Factor and Delayed Retirement Credit Calculation}
\usage{
rf_and_drc(claim_age, nra, rf1, rf2, drc, max_drc_age = 70)
}
\arguments{
\item{claim_age}{Numeric value that represents the age at which a worker first claims benefits}

\item{nra}{Numeric value that represents a worker's Normal Retirement Age}

\item{rf1}{Numeric value that represents the incremental reduction in benefits for the first 36 months prior to the NRA based on the worker's birth cohort.}

\item{rf2}{Numeric value that represents the incremental reduction in benefits for the additional months past 36 that in which benefits are claimed early.}

\item{drc}{Numeric value that represents the incremental increase in benefits for the months claimed past the NRA, based on the worker's birth cohort.}

\item{max_drc_age}{Numeric value for maximum age at which DRCs accrue (default 70).
DRC months = (max_drc_age - nra) * 12. Per 42 USC 402(w).}
}
\value{
act_factor numeric value used for adjusting a worker's PIA to compute their monthly benefit
}
\description{
Function that computes a worker's actuarial adjustment to their benefits based on their claiming age.
}
