% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benefit_calculations.R
\name{worker_benefit}
\alias{worker_benefit}
\title{Retired Worker Benefit Calculation}
\usage{
worker_benefit(worker, assumptions, debugg = FALSE)
}
\arguments{
\item{worker}{Dataframe with a worker's COLA-adjusted retired worker PIA by age}

\item{assumptions}{Dataframe with the Social Security Trustees historical and projected economic variables and program parameters}

\item{debugg}{Boolean value that directs function to output additional variables if set to true}
}
\value{
worker Dataframe with a workjer's retired worker benefit by age
}
\description{
Function that calculates a worker's retirement benefit based on their own earnings record
}
