% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eligibility.R
\name{years_of_coverage}
\alias{years_of_coverage}
\title{Years of Coverage Computation}
\usage{
years_of_coverage(worker, debugg = FALSE)
}
\arguments{
\item{worker}{Data frame with workers' earnings by age/year (must include yoc_threshold column)}

\item{debugg}{Optional boolean variable that allows for the output of additional variables for testing.}
}
\value{
worker Data frame with years_of_coverage column appended
}
\description{
Function for computing a worker's years of coverage for special minimum PIA.
Per 42 USC 415(a)(1)(C)(i), special minimum PIA requires at least 11 years of coverage.
}
\details{
Years of coverage threshold (from SSA OACT https://www.ssa.gov/OACT/COLA/yoc.html):
\itemize{
\item 1951-1978: 25\% of contribution and benefit base
\item 1979-1990: 25\% of old-law contribution base
\item 1991+: 15\% of old-law contribution base
}
}
