% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benefit_calculations.R
\name{final_benefit}
\alias{final_benefit}
\title{Final Benefit Calculation}
\usage{
final_benefit(worker, debugg = FALSE)
}
\arguments{
\item{worker}{Dataframe with a worker's retired worker and spousal benefit by age}

\item{debugg}{Boolean value that directs function to output additional variables if set to true}
}
\value{
worker Dataframe with a worker's final monthly benefit by age
}
\description{
Function that calculates a worker's total retirement benefit combining
retired worker and spousal benefits.
}
