% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptions_prep.R
\name{prep_assumptions}
\alias{prep_assumptions}
\title{Prepare Program Assumptions}
\usage{
prep_assumptions(dataset)
}
\arguments{
\item{dataset}{Dataframe with the latest Social Security Trustees Report assumptions.}
}
\value{
assume Dataframe with TR assumptions and projected program parameters.
}
\description{
Function that imports the Social Security Trustees Report's assumptions and
sets and projections Social Security's main parameters.
}
\examples{
\dontrun{
tr2025 <- prep_assumptions(tr2025)
}

}
